Split test_model.py (3169 LoC) into separate files
